fixed parquet cannot write custom exposed type
